refactor(teacher): split CrossTextAnalytics.tsx 537L into 4 focused modules (Related to #1951)
Extract:
- useCrossTextData hook (hooks/useCrossTextData.ts) — data fetching + error state
- CrossTextFilterBar (components/CrossTextFilterBar.tsx) — view mode toggle
- CrossTextChartGrid (components/CrossTextChartGrid.tsx) — ClassOverviewPanel, StudentDetailPanel, StudentSelector
- CrossTextAnalytics (orchestrator, ~80L vs 537L original)

TDD: 15 characterization tests, all GREEN pre/post refactor.
